+++ tests/Application/Query/Happening/Webinar/GetWebinarViewQueryHandlerTest.php
@@ -77,6 +77,7 @@ class GetWebinarViewQueryHandlerTest extends TestCase
         ];
 
         $happening = $this->prophesize(Happening::class);
+        $happening->getId()->shouldBeCalled()->willReturn(1);
         $happening->getTitle('en')->shouldBeCalled()->willReturn(
             'Webinar: how to work remotely during the Covid-19 crisis'
         );
@@ -124,6 +125,7 @@ class GetWebinarViewQueryHandlerTest extends TestCase
 
         $this->assertEquals(
             new WebinarView(
+                1,
                 111,
                 'Webinar: how to work remotely during the Covid-19 crisis',
                 'User token',