+++ CHANGELOG.md
@@ -5,6 +5,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- VIMEET-2141 - Webinar questions
 - VIMEET-2102 - Evaluate visio meeting
 - VIMEET-2128 - Show participant name on visio